add first hydra tests
